made plugin code compatible for 1.6.x or newer
As $(..).attr() has been deprecated since jquery 1.6 and new function
prop() has been introduced. This commit removes usage of attr() and uses
prop() in favor of continuing support for latest jquery versions.
Though there was no issues reported or we did not found any problem with
attr() but using prop() was a move for keeping it updated.